+++ src/main/java/io/github/langqi99/aeallpattern/mixin/client/ClientJeiAggregateScannerMixin.java
@@ -56,7 +56,7 @@ public abstract class ClientJeiAggregateScannerMixin {
         }
     }
 
-    @WrapOperation(method = "onRightClickBlock", at = @At(value = "INVOKE",
+    @WrapOperation(method = "dispatchScan", at = @At(value = "INVOKE",
             target = "Lio/github/langqi99/aeallpattern/client/ClientJeiAggregateScanner;startScan(Lmezz/jei/api/runtime/IJeiRuntime;Lnet/minecraft/core/BlockPos;)V"))
     private static void preferEmi(IJeiRuntime runtime, BlockPos pos, Operation<Void> original) {
         if (ModList.get().isLoaded("toomanyrecipeviewers")) {